task

+++ Lab-11/KenaKata/src/main/java/kenakata/order/PriceBreakdown.java
@@ -1,4 +1,4 @@
 package kenakata.order;
 
-public record PriceBreakdown(long subtotal,double discount,double vat,long delivery,double insurance,double serviceFee,double grandTotal) {
+public record PriceBreakdown(long subtotal,double discount,double vat,long delivery,double insurance,double serviceFee,long grandTotal) {
 }
